% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_rcmd.R
\name{cxlib_rcmd}
\alias{cxlib_rcmd}
\title{Utility function to run an R program}
\usage{
cxlib_rcmd(
  x,
  log = NULL,
  wd = base::getwd(),
  libpaths = .libPaths(),
  environ = NULL,
  wait = FALSE
)
}
\arguments{
\item{x}{Program path}

\item{log}{Alternative log path}

\item{wd}{Working directory}

\item{libpaths}{Vector of library paths}

\item{environ}{Custom environmental file}

\item{wait}{Wait for program to finish executing}
}
\description{
Utility function to run an R program
}
