% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_actions.R
\name{cxlib_remote_actions}
\alias{cxlib_remote_actions}
\title{(Experimental) Utility function to list actions for a job}
\usage{
cxlib_remote_actions(x, queue = NULL)
}
\arguments{
\item{x}{Job reference}

\item{queue}{URL to R Compute Service}
}
\value{
Invisible job reference
}
\description{
(Experimental) Utility function to list actions for a job
}
