% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_addtoken.R
\name{cxlib_remote_addtoken}
\alias{cxlib_remote_addtoken}
\title{(Experimental) Utility function to store R Compute Service personal access token}
\usage{
cxlib_remote_addtoken()
}
\value{
Invisible NULL
}
\description{
The R Compute Service (rcx) utilities access tokens to identify a permitted
requester. The access token is stored in the users home directory.
}
