% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_delete.R
\name{cxlib_remote_delete}
\alias{cxlib_remote_delete}
\title{(Experimental) Delete a job entry in the remote processing node}
\usage{
cxlib_remote_delete(x, queue = NULL)
}
\arguments{
\item{x}{Job reference}

\item{queue}{URL to R Compute Service}
}
\value{
Invisible job reference
}
\description{
(Experimental) Delete a job entry in the remote processing node
}
