% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_jobinfo.R
\name{cxlib_remote_jobinfo}
\alias{cxlib_remote_jobinfo}
\title{(Experimental) Utility function presenting job information}
\usage{
cxlib_remote_jobinfo(x)
}
\arguments{
\item{x}{Job reference}
}
\value{
Invisible job reference
}
\description{
The utility function displays information about the job
}
