% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_listjobs.R
\name{cxlib_remote_listjobs}
\alias{cxlib_remote_listjobs}
\title{(Experimental) Utility function to list jobs in a remote queue}
\usage{
cxlib_remote_listjobs(queue = NULL)
}
\arguments{
\item{queue}{URL to R Compute Service}
}
\value{
Invisible list of jobs
}
\description{
The function will retrieve the list of registered jobs in the process queue
}
