% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_save.R
\name{cxlib_remote_save}
\alias{cxlib_remote_save}
\title{(Experimental) Utility function to retrieve results from a job}
\usage{
cxlib_remote_save(x, out = base::getwd(), queue = NULL)
}
\arguments{
\item{x}{Job reference}

\item{out}{Output directory path}

\item{queue}{URL to R Compute Service}
}
\value{
Invisible job reference
}
\description{
(Experimental) Utility function to retrieve results from a job
}
