% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_remote_wait.R
\name{cxlib_remote_wait}
\alias{cxlib_remote_wait}
\title{(Experimental) Utility function to induce a timeout period waiting for results}
\usage{
cxlib_remote_wait(x, wait = 5, queue = NULL)
}
\arguments{
\item{x}{Job reference}

\item{wait}{Duration in minutes}

\item{queue}{URL to R Compute Service}
}
\value{
Invisible job reference
}
\description{
A simple utility function to wait for job execution to complete

The \code{wait} duration in minutes. If the job has not completed before the
duration is exceeded,
}
