% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_vault.R
\name{cxlib_vault}
\alias{cxlib_vault}
\title{Connect to a key vault}
\usage{
cxlib_vault()
}
\value{
A standard reference class to access vault secrets
}
\description{
This function returns a reference class object the represents the vault.

\strong{\emph{It is on the role of the developer and implementer to ensure
that the secrets stored in the vault and their use are appropriately protected.}}

The vault service is defined using the cxlib property \code{VAULT}. If the property
is not defined, the request results in an error.

A local vault can be enabled with \code{VAULT} equal to \code{LOCAL}. See
\link{.cxlib_vaultlocal} for configuration options.

A connection to Azure Key Vault can be enabled with \code{VAULT} equal to
\code{AZUREKV}. See \link{.cxlib_vaultazurekv} for configuration options.
}
