% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_vaultlocal.R
\docType{class}
\name{cxlib_vault_local-class}
\alias{cxlib_vault_local-class}
\alias{.cxlib_vaultlocal}
\title{Internal reference class representing a vault on the local file system}
\description{
An internal utility class representing a simple local file system vault to
store secrets.

The vault is hierarchical and uses a path notation with forward slash as
separator and leading slash representing the root of the vault.

The levels of the hierarchy and the secret name consists of the characters
a-z, 0-9 and punctuation dash (-), underscore '_' and period '.'.

The vault storage directory is defined by the cxlib property \code{VAULT.DATA}.

\strong{\emph{It is on the role of the developer and implementer to ensure
that the secrets stored in the vault and their use are appropriately protected.}}

The \code{list()} provides a means to list the secrets stored within the vault.

The \code{secret()} method retrieves a specified secret from the vault. If the
secret is not defined, the value of \code{unset} is returned. A secret is stored and
returned as a single text value.
}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initialize vault}

\item{\code{list()}}{List secrets}

\item{\code{secret(x, unset = NA)}}{Get secret}
}}

\keyword{internal}
