% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_batchjob_actions.R
\name{.cxlib_batchjob_actions}
\alias{.cxlib_batchjob_actions}
\title{Internal function to list actions}
\usage{
.cxlib_batchjob_actions(x)
}
\arguments{
\item{x}{Directory path with actions}
}
\value{
List of actions
}
\description{
Internal function to list actions
}
\keyword{internal}
