% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_batchjob_execd_program.R
\name{.cxlib_batchjob_execd_program}
\alias{.cxlib_batchjob_execd_program}
\title{Internal utility to execute a program as part of batch job actions}
\usage{
.cxlib_batchjob_execd_program(x, options)
}
\arguments{
\item{x}{Action}

\item{options}{Execution options}
}
\value{
A list of named entries representing the action results
}
\description{
The routine takes two lists as input.

All paths are relative to the \code{work.area} specified in the \code{options} parameter.

The action \code{x} is defined as a nested list with the following named elements

\itemize{
\item \code{type} is the action type and equal to \code{program}, case in-sensitive
\item \code{id} identifies the action with an action ID
\item \code{path} the program path
\item \code{sha1} the program SHA-1 hash
\item \code{log} a list that contains the log file path (\code{path})
}

The action list of options \code{options} is defined as a nested list with the
following named elements

\itemize{
\item \code{job.id} identifies the action parent Job ID
\item \code{work.area} represents the working directory for executing the program
}

The result returned is the action list with the additional named elements and updates.

\itemize{
\item \code{log} from the action with the addition of the log file SHA-1 hash (\code{sha1})
\item \code{files.input} a list of input files in the work area at the start of execution
\item \code{files.created} new files created in the work area during execution
\item \code{files.updated} existing files updated in the work area during execution
\item \code{files.deleted} files that existed and was deleted in the work area during execution
}

The entries of \code{files.input}, \code{files.created}, \code{files.updated} and \code{files.deleted} consist
of both the path \code{path} and SHA-1 hash \code{sha1}. For deleted files, the SHA-1 hash represents
the SHA-1 hash of the file before it was deleted.
}
\keyword{internal}
