% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_batchjob_resultsarchive.R
\name{.cxlib_batchjob_resultsarchive}
\alias{.cxlib_batchjob_resultsarchive}
\title{Internal function to generate a results archive from completed actions}
\usage{
.cxlib_batchjob_resultsarchive(x, job.path = NULL, work.path = NULL)
}
\arguments{
\item{x}{Result archive file}

\item{job.path}{Path to job actions}

\item{work.path}{Path to job work area}
}
\value{
Path to archive file
}
\description{
Internal function to generate a results archive from completed actions
}
