% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_rcmd.R
\name{.cxlib_rcmd}
\alias{.cxlib_rcmd}
\title{Internal utility function to run an R program as a background process}
\usage{
.cxlib_rcmd(x, log = NULL, wd = base::getwd())
}
\arguments{
\item{x}{Program path}

\item{log}{Alternative log path}

\item{wd}{Working directory}
}
\value{
Invisible NULL
}
\description{
A crude function to execute an R program as a standalone process

The .Renviron configuration can be coerced using the cxlib property
\code{RENVIRON.DEFAULT} to define standard R compute environment configuration.
}
\keyword{internal}
