% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_remote_accesstoken.R
\name{.cxlib_remote_accesstoken}
\alias{.cxlib_remote_accesstoken}
\title{(Experimental) Internal function to read R Compute Service token}
\usage{
.cxlib_remote_accesstoken()
}
\value{
Invisible character value
}
\description{
(Experimental) Internal function to read R Compute Service token
}
\keyword{internal}
