% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_batchjob.R
\docType{class}
\name{cxlib_batchjob-class}
\alias{cxlib_batchjob-class}
\alias{cxlib_batchjob}
\title{A reference class representing a cxlib batch job}
\description{
A utility class representing a cxlib batch job. A job can be identified by
an existing Job ID or a vector of program paths or a job definition to execute.

A batch job consists of a set of actions, which in the current implementation
is limited to an ordered sequence of one or more programs. The programs can
be specified as a character vector or as the named list entry \code{programs} as
part of the job definition.

\emph{Note: All paths, including program annotations, are relative to the
current working directory.}

If the \code{cxlib_batchjob} is initialized with a single character string in the
format of a valid Job ID, it is assumed the string represents an existing Job.

A vector of program paths can be specified as input. The program paths in the
specified order becomes the new Job definition.

A job definition provided as input is a list of named entries, with only \code{programs}
required.

\itemize{
\item \code{programs} a vector of program paths in the order of execution
\item \code{id} a predefined Job ID in UUID format (see \link[uuid]{UUIDgenerate})
\item \code{options} sets one or more options (see below) for the submitted Job

}

The following options are supported as part of a Job definition or as a cxlib
property (see process options below).

\itemize{
\item \code{LOGS} defines the standard directory path for logs
\item \code{LOG.FILEEXT} defines the standard log file extension
}

If the option is defined as both part of the Job and a property, the job
definition takes precedence.

Processing options can be defined as part of a Job definition or through cxlib
properties (see \link{cxlib_config}).

The following options sets the locations of for processing.
\itemize{
\item \code{PATH}option specifies the root directory for batch job processes. If
not defined, the sub-directory \code{.cxlib} in the R session temporary directory
(\link[base]{tempdir}) is used.
\item \code{WORKPATH} option specifies the work area directory for batch jobs.
If not defined, the directory \code{PATH} is used. The \code{WORKPATH} is the
parent directory to job working directories.
}

The method \code{sumbit()} submits the registered actions for processing. Actions
are executed in a standalone R session. If \code{WAIT} is equal to \code{FALSE}, the
R session is executed in the background.

The \code{save()} methods saves the action results to the current working directory.

The \code{delete()} method deletes the directories associated with the job.

The experimental method \code{actions()} returns a list of actions and their status.

The experimental method \code{status()} returns a list of actions with abbreviated
information including their status.

The experimental method \code{details()} returns a list of job details including
job actions and status.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage}
}}

\section{Methods}{

\describe{
\item{\code{actions()}}{List actions}

\item{\code{delete()}}{Delete job areas}

\item{\code{details(wait = TRUE)}}{Job details}

\item{\code{initialize(x)}}{Create new batch job}

\item{\code{save()}}{Save job results}

\item{\code{status()}}{Job status}

\item{\code{submit(wait = TRUE)}}{Submit job}
}}

